Committing old uncommitted changes from months ago. Most appear to be refactoring or protocol tweaking.

+++ src/ca/formulahybrid/telemetry/message/ControlFlagDescriptor.java
@@ -11,6 +11,6 @@ import java.lang.annotation.Target;
 @Documented
 public @interface ControlFlagDescriptor {
 
-	short id();
+	byte id();
 	int length();
 }
